Fixed breakage in the image class. Needed to use 16-bit channels instead of 8-bit.

+++ inc/ed_pixel.h
@@ -3,9 +3,9 @@
 #define _ED_PIXEL_H_
 
 struct pixel_t {
-    uint8_t red;
-    uint8_t green;
-    uint8_t blue;
+    uint16_t red;
+    uint16_t green;
+    uint16_t blue;
 
     /* Overloaded operators for comparing pixels */
     // TODO re-evaluate only using red channel here
